# services/gateway/swagger.yaml
openapi: 3.0.0
info:
  title: API Gateway
  description: API Gateway for microservices
  version: 1.0.0
servers:
  - url: http://localhost:2000
    description: Local development server
paths:
...
  /api/target/photo:
    delete:
      summary: Delete a photo by ID (user must be the owner)
      parameters:
        - in: path
          name: photoId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Photo deleted successfully
        "403":
          description: Unauthorized to delete
        "404":
          description: Photo not found
        "500":
          description: Internal server error

  ####################################################
  # Read Service
  ####################################################
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT